[BUGFIX] Internally store boolean properties as integers (#360)
This avoids problems when persisting models with the ConnectionPool.

+++ Classes/Object.php
@@ -212,7 +212,7 @@ abstract class Tx_Oelib_Object
     {
         $this->checkForNonEmptyKey($key);
 
-        $this->set($key, (bool)$value);
+        $this->set($key, (int)(bool)$value);
     }
 
     /**
